refactor: feign client 패키지 명을 client로 변경 - feign client, web client, rest client 존재하기 때문
