+++ src/main/java/com/example/tripmingle/client/KakaoLoginFeignClientPort.java
@@ -1,4 +1,4 @@
-package com.example.tripmingle.feign;
+package com.example.tripmingle.client;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
